model deployment/requirements.txt
-i https://pypi.org/simple
aiosqlite==0.21.0; python_version >= '3.9'
alembic==1.16.5; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.10.0; python_version >= '3.9'
apprise==1.9.4; python_version >= '3.9'
asgi-lifespan==2.1.0; python_version >= '3.7'
asyncpg==0.30.0; python_full_version >= '3.8.0'
attrs==25.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.8.3; python_version >= '3.7'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
cloudpickle==3.1.1; python_version >= '3.8'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
coolname==2.2.0
cryptography==45.0.7; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
dateparser==1.2.2; python_version >= '3.8'
docker==7.1.0; python_version >= '3.8'
exceptiongroup==1.3.0; python_version >= '3.7'
fastapi==0.116.1; python_version >= '3.8'
flask==3.1.2; python_version >= '3.9'
fsspec==2025.7.0; python_version >= '3.9'
google-api-core==2.25.1; python_version >= '3.7'
google-auth==2.40.3; python_version >= '3.7'
google-cloud-core==2.4.3; python_version >= '3.7'
google-cloud-storage==3.3.1; python_version >= '3.7'
google-crc32c==1.7.1; python_version >= '3.9'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos==1.70.0; python_version >= '3.7'
graphviz==0.21; python_version >= '3.9'
greenlet==3.2.4; python_version < '3.14' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
griffe==1.13.0; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
h2==4.3.0; python_version >= '3.9'
hpack==4.1.0; python_version >= '3.9'
httpcore==1.0.9; python_version >= '3.8'
httpx[http2]==0.28.1; python_version >= '3.8'
humanize==4.13.0; python_version >= '3.9'
hyperframe==6.1.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.7.0; python_version >= '3.9'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jinja2-humanize-extension==0.4.0; python_version >= '3.0'
joblib==1.5.2; python_version >= '3.9'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.25.1; python_version >= '3.9'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
mako==1.3.10; python_version >= '3.8'
markdown==3.8.2; python_version >= '3.9'
markdown-it-py==4.0.0; python_version >= '3.10'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
numpy==2.3.2; python_version >= '3.11'
oauthlib==3.3.1; python_version >= '3.8'
opentelemetry-api==1.36.0; python_version >= '3.9'
orjson==3.11.3; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
pandas==2.3.2; python_version >= '3.9'
pathspec==0.12.1; python_version >= '3.8'
pendulum==3.1.0; python_version < '3.13'
prefect==3.4.15; python_version < '3.14' and python_version >= '3.9'
prometheus-client==0.22.1; python_version >= '3.9'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==6.32.0; python_version >= '3.9'
pyarrow==21.0.0; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-extra-types==2.10.5; python_version >= '3.8'
pydantic-settings==2.10.1; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.1.1; python_version >= '3.9'
python-slugify==8.0.4; python_version >= '3.7'
python-socks[asyncio]==2.7.2; python_full_version >= '3.8.0'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
readchar==4.2.1; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
regex==2025.8.29; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
requests-oauthlib==2.0.0; python_version >= '3.4'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rich==14.1.0; python_full_version >= '3.8.0'
rpds-py==0.27.1; python_version >= '3.9'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
ruamel-yaml==0.18.15; python_version >= '3.8'
ruamel.yaml.clib==0.2.12; python_version < '3.14' and platform_python_implementation == 'CPython'
scikit-learn==1.5.0; python_version >= '3.9'
scipy==1.16.1; python_version >= '3.11'
semver==3.0.4; python_version >= '3.7'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy[asyncio]==2.0.43; python_version >= '3.7'
starlette==0.47.3; python_version >= '3.9'
text-unidecode==1.3
threadpoolctl==3.6.0; python_version >= '3.9'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typer==0.16.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
tzlocal==5.3.1; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uv==0.8.14; python_version >= '3.8'
uvicorn==0.35.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
zipp==3.23.0; python_version >= '3.9'
